Command

Description

Module in charge of parsing the CLI command and constructing a domain specific representation of the instruction to execute.

Documentation

data GenCommand

GenCommand represents the instructions the program will execute. It contains what we are going to copy, its name and content substitutions as well as whether or not to treat the artifact as a module.

Constructors

GenCommand

Fields

what :: String

What do we want to copy, or which template do we want to use.

name :: String

The name that will be given to the new files.

sub :: Map String String

Mapping of values to substitute in the templates content

asModule :: Bool

Should the output files be treated as a module. If it is a module it will create a directory with the name and put the files inside.

output :: Maybe RelDir

Alternate output directory. Use to override config parameter

data CommandError

Encompasses all possible errors in the Command module

Constructors

InvalidSubstitutionError String
If substitution does is not a key/value pair or it contains empty strings

DuplicateSubstitutionError String
If more than one value has been passed for the same substitution key.

toListOfStr :: String -> [[String]]

This function is meant to split the str by ',' first and then split each result by : and then trim each string in the inner List. The IDE helped format it this way but it's a tad hard to grok. e.g. if str is "$A$:a, $B$:b" then the result is [[A, "a"], [B, "b"]]

hasEmptyStrings :: [String] -> Bool

mkPair :: [String] -> Either CommandError (String, String)

safeInsert :: (String, String) -> Map String String -> Either CommandError (Map String String)

mkSubstitutions :: [[String]] -> Either String (Map String String)

mkOptionalOutput :: String -> Either String (Maybe RelDir)

mkGenCommand

Arguments

:: String
What do we want to copy, or which template do we want to use.

-> String
The name that will be given to the new files.

-> Map String String
Mapping of values to substitute in the templates content.

-> Bool
Should the output files be treated as a module. If it is a module it will create a directory with the name and put the files inside.

-> Maybe RelDir
Alternate output directory. Use to override config parameter

-> GenCommand

GenCommand factory

mkGenCommandParser :: Parser GenCommand

parserOptions :: ParserInfo GenCommand

This is the configuration given to Opt-Parser Applicative. This is what will be used in the cli program, the messages and banners displayed. It is a parser for GenCommand.