Config

Description

Contains metadata for the application as well as values needed for Progen to run.

Documentation

data Dotfile

It represents the data coming from the configuration file. More data can be needed but may not be configurable. The Dotfile is used to create a GenConfig which will be used by the application.

Constructors

Dotfile

Fields

templates :: Last RelDir

Templates directory path relative to the root.

output :: Last (Map String RelDir)

Configuration for the output. Keys are matched on the names of the templates, the values are relative directory paths inside the root.

filenameSeparator :: Last (Maybe Char)

Character used to separate the template filename and make use of suffix.

newtype ConfigError

Encompasses all possible errors in the Config module

Constructors

CouldNotMakeConfig String
Cannot create a valid config from all inputs

handleConfigResult :: Maybe GenConfig -> Either ConfigError GenConfig

Transform maybe config to an Either of the correct Domain error or the valid Config

mkDirPath :: (String -> Dir os) -> Text -> Last (Dir os)

mkOutputDirs :: Map String String -> Last (Map String RelDir)

data GenConfig

This is the configuration used by the application wtih all the metadata needed to generate code.

Constructors

GenConfig

Fields

templatesDir :: RelDir

Templates directory path relative to the root.

outputDirs :: Map String RelDir

Configuration for the output. Keys are matched on the names of the templates, the values are relative directory paths inside the root.

separator :: Maybe Char

Character used to separate the template filename and make use of suffix.

dotfileName :: String

Name for the dotfile where the configuration will be read from

makeDefaultConfig :: Dotfile

emptyConfigOption :: Dotfile

decodeConfig :: String -> Dotfile

mkDotfile

Arguments

:: Last RelDir
Templates directory path relative to the root.

-> Last (Map String RelDir)
Configuration for the output. Keys are matched on the names of the templates, the values are relative directory paths inside the root.

-> Last (Maybe Char)
Character used to separate the template filename and make use of suffix.

-> Dotfile

Factory function for constructing a Dotfile

mkConfig :: Dotfile -> String -> Maybe GenConfig

Factory to make a GenConfig from a JSONString. It will attempt to construct a Dotfile from its values and then combine it with a default config. The result will be the GenConfig that will be returned.

>>> mkConfig "{ \"root\": \"/dummy/project\", \"templates\": \".progen/templates\", \"filenameSeparator\": \".\", \"output\": { \"component\": \"./components\" }}"